+++ lib/dart_ytmusic_api.dart
@@ -5,6 +5,7 @@ export './enums.dart';
 export './types.dart';
 export './utils/artists.dart';
 export './utils/traverse.dart';
+export './utils/playable_video_id.dart';
 export './parsers/album_parser.dart';
 export './parsers/artist_parser.dart';
 export './parsers/parser.dart';